requirements.txt
astroplan
astroquery
matplotlib
pyyaml
